requirements.txt
# general
paramiko==2.0.2
pytest==3.0.4
pytest-xdist==1.15.0
python-cinderclient==1.9.0
python-glanceclient==2.5.0
python-heatclient==1.5.0
python-keystoneclient==3.5.0
python-muranoclient==0.11.0
python-neutronclient==6.0.0
python-novaclient==6.0.0
python-openstackclient==3.3.0
python-swiftclient==3.1.0
requests==2.11.1
six==1.10.0
waiting==1.3.0
subprocess32==3.2.7; python_version < '3.2' and os_name == 'posix'
mock==2.0.0
pyaml==16.9.0
python-dateutil==2.5.3
prettytable==0.7.2
dpkt==1.8.8
os-faults>=0.1.7,<0.2.0
attrdict==2.0.0
PyHamcrest==1.9.0

# debug
ipdb==0.10.1
ipython==5.1.0

# doc
Sphinx==1.4.6
oslosphinx==4.7.0

# verification
tox==2.4.1
